Added support for the superfast hash, and added a configurable number of bins.
After some profiling, I found that much of the time spent in these operations
was in the hashing function. Replacing md5 with superfast decreased times in
the speedtest by a factor of 3 or 4, and didn't reduce accuracy. MD5 was simply
overkill as a hashing function as we really just need a deterministic,
randomizing function.